feat: extract watch logic as Node.js API
