+++ lib/api.js
@@ -1,6 +1,7 @@
 const { PassThrough } = require('stream')
 const { build } = require('./build')
 const { dev: startDevServer } = require('./dev')
+const { watch } = require('./watch')
 
 function isDev () {
   return process.env.NODE_ENV !== 'production'
@@ -22,3 +23,4 @@ exports.dev = dev
 exports.prod = prod
 exports.build = build
 exports.startDevServer = startDevServer
+exports.watch = watch
