+++ lib/build.js
@@ -1,3 +1,5 @@
+'use strict'
+
 const vfs = require('vinyl-fs')
 const hashSum = require('hash-sum')
 const loadConfig = require('./config').loadConfig
